Clears compiler warning by changing type

+++ src/StringTokenizer.cpp
@@ -10,7 +10,7 @@
 #include "StringTokenizer.h"
 
 
-void Tokenize(const unsigned char* STR, size_t szStr, vector<string> &tokens, vector<PSInt> &tokenSizes, size_t *szTotalTokens, const char* DELIMITERS, int szDELIMITERS)
+void Tokenize(const unsigned char* STR, size_t szStr, vector<string> &tokens, vector<PSInt> &tokenSizes, size_t *szTotalTokens, const char* DELIMITERS, size_t szDELIMITERS)
 {
     // Skip delimiters at beginning.
 	string str((const char*)STR, szStr);
